Int64 -> Int

+++ src/network/vanilla_hash_table.jl
@@ -1,4 +1,4 @@
-const Id = Int64
+const Id = Int
 
 mutable struct HashTable
     buckets::Vector{Vector{Id}}
